refactor(toast): move processor-rejection toast routing to the app layer (#142)
`utils/processorRejectionToast.ts` and `utils/AliasCollisionToast.tsx`
imported `@/plugins/alias/collisionMerge` — a `utils` → `plugins`
dependency inversion (the leaf util layer depending on a plugin).

The rejection router is inherently plugin-aware (it renders the alias
plugin's collision-merge toast), so it belongs in the app layer where
depending on `@/plugins/*` is the established direction
(`staticAppExtensions.ts` already imports every plugin). Relocate both
files to `src/extensions/`, fix the `@/utils/toast` import, and update
the wiring import in `context/repo.tsx`. No behavior change.
